feat(auth): implement forgot password screen and functionality
feat(auth): add login screen and related DTOs

feat(auth): create signup screen and necessary DTOs

feat(auth): add email verification screen

feat(cart): create empty cart screen

feat(home): implement home screen layout

feat(onboarding): add onboarding screens with animations

feat(order): implement order creation and update status DTOs

feat(payment): add payment processing DTOs for Momo and Stripe

feat(point): create point exchange and voucher status update DTOs

feat(product): implement slug DTO for product identification

feat(refund): add refund processing DTOs for Momo and Stripe

feat(review): create review creation and update DTOs

feat(shipping): implement shipping fee calculation DTO

feat(wishlist): create favorite screen for user wishlist

feat(routing): implement main navigation with bottom navigation bar

chore(utils): add date formatting utility
