+++ lib/src/features/home/presentation/home_screen.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:techbox/src/routing/main_navigation.dart';
+
+class HomeScreen extends StatefulWidget {
+  const HomeScreen({super.key});
+
+  @override
+  State<HomeScreen> createState() => _MyWidgetState();
+}
+
+class _MyWidgetState extends State<HomeScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Đây là trang home'),
+        backgroundColor: Colors.white,
+      ),
+    );
+  }
+}
