+++ lib/src/features/product/data/slug_dto.dart
